Bringing fork up to date

+++ automatic/thunderbird/tools/chocolateyInstall.ps1
@@ -1,4 +1,4 @@
-$ErrorActionPreference = 'Stop'
+﻿$ErrorActionPreference = 'Stop'
 # This is the general install script for Mozilla products (Firefox and Thunderbird).
 # This file must be identical for all Choco packages for Mozilla products in this repository.
 $toolsPath = Split-Path $MyInvocation.MyCommand.Definition
@@ -7,19 +7,17 @@ $toolsPath = Split-Path $MyInvocation.MyCommand.Definition
 $packageName = 'thunderbird'
 $softwareName = 'Mozilla Thunderbird'
 
-$allLocalesListURL = 'https://www.mozilla.org/en-US/thunderbird/all/'
+$alreadyInstalled = (AlreadyInstalled -product $softwareName -version '52.2.1')
 
-$alreadyInstalled = (AlreadyInstalled -product $softwareName -version '45.5.1')
-
-if ($alreadyInstalled) {
+if ($alreadyInstalled -and ($env:ChocolateyForce -ne $true)) {
   Write-Output $(
     "Thunderbird is already installed. " +
     'No need to download an re-install again.'
   )
 } else {
 
-  $locale = GetLocale -localeUrl $allLocalesListURL -product $softwareName
-  $checksums = GetChecksums -language $locale -checksumFile "$toolsPath\LanguageChecksums"
+  $locale = GetLocale -localeFile "$toolsPath\LanguageChecksums.csv" -product $softwareName
+  $checksums = GetChecksums -language $locale -checksumFile "$toolsPath\LanguageChecksums.csv"
 
   $packageArgs = @{
     packageName = $packageName
@@ -28,7 +26,7 @@ if ($alreadyInstalled) {
 
     Checksum = $checksums.Win32
     ChecksumType = 'sha512'
-    Url = "https://download.mozilla.org/?product=thunderbird-45.5.1-SSL&amp;os=win&amp;lang=${locale}"
+    Url = "https://download.mozilla.org/?product=thunderbird-52.2.1-SSL&os=win&lang=${locale}"
 
     silentArgs = '-ms'
     validExitCodes = @(0)
